Implement IDamageable. Implement AttackSystem.
Fix minorbugs in AttackState
- fix not attack on time.

+++ Assets/Game/BattleCore/Status/Status.cs
@@ -5,8 +5,8 @@ public class Status
 {
     public int hp;
     public int maxHp;
-    public int attack;
-    public int defense;
+    public float attack;
+    public float defense;
     public float speed;
 
     public Status()
